+++ hutool-core/src/main/java/cn/hutool/core/lang/mutable/MutableFloat.java
@@ -46,7 +46,7 @@ public class MutableFloat extends Number implements Comparable<MutableFloat>, Mu
 
 	@Override
 	public Float get() {
-		return Float.valueOf(this.value);
+		return this.value;
 	}
 
 	/**
